formatter/requirements.txt
streamlit>=1.30.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
spacy>=3.7.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
httpx>=0.27.0
